+++ test/svs/core_test.go
@@ -38,6 +38,6 @@ func TestCoreInitialState(t *testing.T) {
 		UpdateCallback: func(missing []svs.MissingData) { return },
 	}
 	core := svs.NewCore(nil, config, svs.GetDefaultConstants())
-	assert.Equal(t, uint(0), core.GetSeqno())
+	assert.Equal(t, uint64(0), core.GetSeqno())
 	assert.Equal(t, svs.NewStateVector(), core.GetStateVector())
 }
